refactor(orzma_tmux): rename command::copymode module -> vi_mode
Keeps PromptKind::copy_command (returns tmux -X protocol strings).
